Allow users to set a remote certificate validation callback for the IoT hub service client
currently supported over HTTP and AMQP TCP. Will add AMQP WS support later
